Move responsiblity for path prompting to the repository
The type of the paths depends on the kind of repository being used in the project.
For file-based repositories like Git, the paths are actually file paths,
so it makes sense to use file or directory pickers.

Use the new file and directory dialogs introduced with Squeak 5.2,
since they have gained an add directory button by now.

This time I tried a continuation-style control flow instead of using cancel blocks.
It avoids using modal dialogs to simulate a "procedural" control flow.
Message passing allows the business process to continue after users finished their
interaction with the interface. The downsides are that you cannot see the stack trace
that lead to the invocation of the UI and that you must not return from the continuation
blocks as their home method contexts are no longer on the stack by the time the block
is invoked.