Prevent the delegate object registry from being serialized with a shadow graph
The registry can contain objects that are not (or no longer) part of the graph
to be captured and stored. Further, the registry is shared among all object graphs
in a SquotImageStore, so the registry were stored with each shadow graph, it would
be redundantly stored for all tracked objects (that serialize the shadow graph object itself).