Fix recognizing of start objects after materialization
Preserve object names from shadow graph even if the materialized
object says it does not need a name. Useful for start objects that
don't usually have names, but get #() assigned to retrieve them
as start objects.

Fix materialize tests for nil, true, and false by wrapping these objects
in a value holder. It makes no sense to track either of these alone
to it is ok not to support that.

Note that a name for nil cannot be remembered because nil can't
be used as the key in a WeakIdentityDictionary used to store the
names for objects in SquotShadowGraph.

(cherry-picked from commit 3d4938affdbc76a9d797fd8527bf2f3613a98472)